All Apis done

+++ api/models/Employee.php
@@ -106,6 +106,10 @@
 
         // Delete Employee
         public function deleteEmployee($id){
+            if($this->getEmployee($id) == false){
+                $this->message = array('message' => 'No employee associated with this id', 'status' => 404);
+                return false;
+            }
             $query = 'DELETE FROM '.$this->table.' WHERE id = ?';
             $stmt = $this->conn->prepare($query);
             $stmt->bind_param('i', $id);
