+++ api/models/Finance.php
@@ -1,16 +1,16 @@
 <?php
-    class Project{
+    class Finance{
 
         // DB stuff
         private $conn;
-        private $table = 'projects';
+        private $table = 'finance';
         public $message = array();
 
-        // Project Properties
+        // Finance Properties
         public $id;
-        public $project_name;
-        public $category;
-        public $location;
+        public $expense_name;
+        public $expense_type;
+        public $amount;
         public $doc;
         public $created;
 
@@ -19,44 +19,44 @@
             $this->conn = $db;
         }
 
-        // Get Projects
-        public function getProjects(){
+        // Get Finances
+        public function getFinances(){
                 $query = 'SELECT * FROM '.$this->table;
                 $stmt = $this->conn->prepare($query);
                 $stmt->execute();
                 $result = $stmt->get_result();
                 if($result->num_rows > 0){
-                    $projects = array();
+                    $finances = array();
                     while($row = $result->fetch_assoc()){
                         $this->id = $row['id'];
-                        $this->project_name = $row['project_name'];
-                        $this->category = $row['category'];
-                        $this->location = $row['location'];
+                        $this->expense_name = $row['expense_name'];
+                        $this->expense_type = $row['expense_type'];
+                        $this->amount = $row['amount'];
                         $this->doc = $row['doc'];
                         $this->created = $row['created'];
-                        $project = $this->toArray();
-                        array_push($projects, $project);
+                        $finance = $this->toArray();
+                        array_push($finances, $finance);
                     }
                     $stmt->close();
                     
                     $this->message = array('message' => 'success', 'status' => 200);
-                    return $project;
+                    return $finance;
                 }
                 $stmt->close();
-                $this->message = array('message' => 'Projects not found', 'status' => 404);
+                $this->message = array('message' => 'Finances not found', 'status' => 404);
                 return false;     
         }
 
-        // Create Project
-        public function createProject(){
-            $query = 'INSERT INTO '.$this->table.'(project_name,category,location,doc) VALUES (?,?,?,?)';
+        // Create Finance
+        public function createFinance(){
+            $query = 'INSERT INTO '.$this->table.'(expense_name,expense_type,amount,doc) VALUES (?,?,?,?)';
             $stmt = $this->conn->prepare($query);
-            $stmt->bind_param('ssss', $this->project_name, $this->category, $this->location,$this->doc);
+            $stmt->bind_param('ssds', $this->expense_name, $this->expense_type, $this->amount,$this->doc);
             if($stmt->execute()){
                 $this->id = $stmt->insert_id;
                 $this->created = date('Y-m-d H:i:s');
                 $stmt->close();
-                $this->message = array('message' => 'Project Created Successfully !', 'status' => 201);
+                $this->message = array('message' => 'Finance Created Successfully !', 'status' => 201);
                 return true;
             }
             $stmt->close();
@@ -64,8 +64,8 @@
             return false;
         }
 
-        //get single project
-        public function getProject($id){
+        //get single finance
+        public function getFinance($id){
             $query = 'SELECT * FROM '.$this->table.' WHERE id = ?';
             $stmt = $this->conn->prepare($query);
             $stmt->bind_param('i', $id);
@@ -75,28 +75,28 @@
                 $row = $result->fetch_assoc();
                 $stmt->close();
                 $this->id = $row['id'];
-                $this->project_name = $row['project_name'];
-                $this->category = $row['category'];
-                $this->location = $row['location'];
+                $this->expense_name = $row['expense_name'];
+                $this->expense_type = $row['expense_type'];
+                $this->amount = $row['amount'];
                 $this->doc = $row['doc'];
                 $this->created = $row['created'];
                 $this->message = array('message' => 'success', 'status' => 200);
                 return $this->toArray();
             }
             $stmt->close();
-            $this->message = array('message' => 'project not found', 'status' => 404);
+            $this->message = array('message' => 'finance not found', 'status' => 404);
             return false;
         }
 
 
-        // Update Project
-        public function updateProject($id){
-            $query = 'UPDATE '.$this->table.' SET project_name = ?, category = ?, location = ?, doc = ? WHERE id = ?';
+        // Update Finance
+        public function updateFinance($id){
+            $query = 'UPDATE '.$this->table.' SET expense_name = ?, expense_type = ?, amount = ?, doc = ? WHERE id = ?';
             $stmt = $this->conn->prepare($query);
-            $stmt->bind_param('ssssi', $this->project_name, $this->category, $this->location, $this->doc, $id);
+            $stmt->bind_param('ssdsi', $this->expense_name, $this->expense_type, $this->amount, $this->doc, $id);
             if($stmt->execute()){
                 $stmt->close();
-                $this->message = array('message' => 'Project Updated Successfully !', 'status' => 200);
+                $this->message = array('message' => 'Finance Updated Successfully !', 'status' => 200);
                 return true;
             }
             $stmt->close();
@@ -104,29 +104,34 @@
             return false;
         }
 
-        // Delete Project
-        public function deleteProject($id){
-            $query = 'DELETE FROM '.$this->table.' WHERE id = ?';
-            $stmt = $this->conn->prepare($query);
-            $stmt->bind_param('i', $id);
-            
-            if($stmt->execute()){
+        // Delete Finance
+        public function deleteFinance($id){
+            if($this->getFinance($id) == false){
+                $this->message = array('message' => 'No finance associated with this id', 'status' => 404);
+                return false;
+            }else{
+                $query = 'DELETE FROM '.$this->table.' WHERE id = ?';
+                $stmt = $this->conn->prepare($query);
+                $stmt->bind_param('i', $id);
+                
+                if($stmt->execute()){
+                    $stmt->close();
+                    $this->message = array('message' => 'Finance Deleted Successfully !', 'status' => 200);
+                    return true;
+                }
                 $stmt->close();
-                $this->message = array('message' => 'Project Deleted Successfully !', 'status' => 200);
-                return true;
+                $this->message = array('message' => 'Something went wrong', 'status' => 500);
+                return false;
             }
-            $stmt->close();
-            $this->message = array('message' => 'Something went wrong', 'status' => 500);
-            return false;
         }
 
         // Convert to array
         public function toArray(){
             return array(
                 'id' => $this->id,
-                'project_name' => $this->project_name,
-                'category' => $this->category,
-                'location' => $this->location,
+                'expense_name' => $this->expense_name,
+                'expense_type' => $this->expense_type,
+                'amount' =>  $this->amount,
                 'doc' => $this->doc,
                 'created' => $this->created
             );
